requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
piper-tts==1.2.0
soundfile==0.12.1
numpy==1.24.3
torch==2.1.0
piper-tts==1.2.0